Support effective tree shaking with ESM (#246)
* Support effective tree shaking with ESM

Webpack was not able to effectively tree shake when everything was
included in one module.  This update makes it easier for webpack to tree
shake.  It _should_ be backwards compatible because we updated the index
for the ESM module in `package.json`.  However, we were packaging our
own version of `React` and `ReactDOM` in the package.  This caused the
bundle size to get quite large.  Moving these to peer dependencies
reduces the size considerably.

Note, I did my best to test this locally with yarn, but I plan on doing
an pre release and testing with that before actually releasing this as a
proper version.

* Update rollup dependencies

Only the babel plugin needed to be updated according to the yarn
versions.

* v9.0.0-alpha.0

* Extend existing externals

* Clear out the side effect that was breaking everything

* v9.0.0-alpha.5

* Fix eslint error