Remove rollup-plugin-size-snapshot
This tool runs a second Rollup build and a webpack build inside of
the Rollup build, using an in memory filesystem. With a moderately
large dist folder this spawns a *lot* of Node processes and quickly
causes an OOM error. It's also very slow.

If we could exclude the cjs bundle it would work again but I don't
think it's worth it to try and fix the lib since we have bundlewatch.