Optimize rollup (#1049)
* Use the babel runtime helpers instead of bundling

This avoids inlining the babel helpers multiple times. The
runtime helpers package is already included by dependencies of
Radiance.

* Build modules to dist/cjs and dist/csm

- Preserve modules
- Use `cjs` and `esm` prefixes for cjs and esm formats respectively

* Use the .cjs extension for cjs modules

This ensures the modules load correctly when they are imported
using ES imports from a node package using ESM mode.

* Build cjs and esm modules to the same directory

* Write types to dist instead of lib

* Only run Babel via Rollup

We run Babel twice: once from src -> lib via Babel CLI and once
from src -> dist via Rollup. The lib build can't be used without
a bundler and the dist build couldn't be used for everything because
it was missing files. We can remedy this by only creating one ES
module build that can be used everywhere without further bundling.
This will ensure we don't end up in a situation where certain module
formats are missing functionality or broken and we don't notice.

* Configure package exports

This allows us to continue using imports such as `radiance-ui/lib/constants`
even though the directory is actually `dist/constants`. In the future we can
use this to limit access to internal modules.

See:

https://nodejs.org/api/packages.html#packages_main_entry_point_export
https://webpack.js.org/guides/package-exports/

* Run TSC via rollup

* Specify named mode for output.exports

This prevents a warning when generating Common JS.

See https://rollupjs.org/guide/en/#outputexports

* Add peer dependency tslib

* Create separate Node ES and browser ES modules

If we don't convert Radiance itself to a ES module then trying to
load an ES module from the radiance package without the .mjs ext
fails. Separate exports for browser and node imports fixes it.

* Fix rollup external checks

Using /node_modules/ doesn't work because it's already been
resolved by plugin-node-resolve at that point.

* Add missing entrypoints

* Fix export path mapping

* Use explicit exports instead of wildcards

Prevents accidentally exposing things we don't want to.

* Consolidate ES module versions

The node version won't work until Emotion is fixed and by then
we will be able to switch to native ES modules too, so we won't
need two versions anymore.

* Export /lib/utils

* Remove rollup-plugin-size-snapshot

This tool runs a second Rollup build and a webpack build inside of
the Rollup build, using an in memory filesystem. With a moderately
large dist folder this spawns a *lot* of Node processes and quickly
causes an OOM error. It's also very slow.

If we could exclude the cjs bundle it would work again but I don't
think it's worth it to try and fix the lib since we have bundlewatch.

* Remove unnecessary yarn-run-all calls

* Use the correct name for @babel/preset-env

I assume this was just being ignored?

* Use { modules: auto } in all environments

This was only necessary before because we were running Babel without
Rollup. Now that we're using @babel/runtime which uses deep imports,
setting { modules: false} was breaking Storybook, which needs to
build our code both for CJS and ESM.

* Switch babel config to json

This doesn't need to be dynamic anymore, Babel strongly recommends
JSON config files, and it makes it easier to switch from CJS to
ESM in the future if we don't have CJS modules in the repo.

* Ensure all .js files are ESM

This makes it easier to switch Radiance to native ESM in the future.

* Remove rollup commonjs and node-resolve plugins

These plugins are only necessary if you want to bundle dependencies
from node_modules. We don't -- everything in node_modules is external.

This makes the build a lot faster and lets us simplify the external
regex (because node-resolve isn't rewriting the import paths)

* Update lib references

* Restore clean call for build:icons:all

* Remove .cjs files from the bundlewatch config

* Change the external regex again

This suppresses a rollup warning.

* Turn on noEmitOnError so the build fails for type errors

* DRY up build steps

* Add comment to clarify rollup external config