Access NODE_ENV with dot operator (#1098)
Using the bracket syntax can cause issues with various tools that
try to statically replace the environment variables like Next.JS
and Vite. The dot syntax is easier to replace when the app is built.

The brackets were used to make TypeScript happy, so we can tell
TypeScript that we expect the value to exist instead.