Add support for React 18 (#1804)
* Update to React 18

* Switch CircleCI images to directly use 16

The CircleCI builds stopped working since we were using the "lts" tag for node, which was 16, but is now 18. 18 isn't compatible, so this change is needed to get things working again.

* Remove PropTypes

Necessary because as of React 18, the types for PropTypes.node no longer match with React.ReactNode. This is likely because PropTypes are old, unmaintained, and mostly unused these days. Removing them is not likely to impact anyone.